Lean pass: single fuel path, drop dead code, clippy-clean all targets
Deduplication and dead-code removal from the code-analysis pass:
- engine.rs delegates to the fueling.rs speed-density pipeline
  instead of a second invented pulse-width formula — one fuel math in
  the crate
- delete the defaults::wiring::sensors alias layer (engine.rs and
  datalog.rs now call analog.rs directly)
- cut speculative firing presets (INLINE_4_*, V_TWIN) and the unused
  TleOutput::default_function; keep only what an engine uses, fold pin
  docs onto the enum variants
- drop the thin rpm_from_period_us re-export from datalog

Clippy (all targets, host + thumbv7em firmware + replay bin): zero
warnings. Fixed same-type u16 casts, is_multiple_of, field-assign
after Default, unused import. base_pulse_ms 8 args -> a
SpeedDensityInputs struct (also clearer at every call site).

Net -23 lines, 95 host tests, firmware size unchanged.

+++ src/tle8888.rs
@@ -17,12 +17,12 @@ pub const CMD_READ: u16 = 0;
 
 /// Build a 16-bit write command: register address + 8-bit payload.
 pub const fn cmd_w(reg: u8, data: u8) -> u16 {
-    (CMD_WRITE | ((reg as u16 & 0x7f) << 1) | ((data as u16) << 8)) as u16
+    CMD_WRITE | ((reg as u16 & 0x7f) << 1) | ((data as u16) << 8)
 }
 
 /// Build a 16-bit read command.
 pub const fn cmd_r(reg: u8) -> u16 {
-    (CMD_READ | ((reg as u16 & 0x7f) << 1)) as u16
+    CMD_READ | ((reg as u16 & 0x7f) << 1)
 }
 
 pub const CMD_SR: u16 = cmd_w(0x1a, 0x03);
